requirements.txt
chromadb==0.6.3
dynaconf==3.2.6
graphrag>=1.2.0
graspologic==3.4.1
jsonschema==4.23.0
langchain==0.3.19
langchain_core==0.3.36
langchain_milvus==0.1.8
langchain_ollama==0.2.3
langfuse==2.59.3
langgraph==0.2.74
minio==7.2.15
networkx==3.4.2
ollama==0.4.7
pandas==2.2.3
pydantic==2.10.6
pymilvus==2.5.4
pymupdf4llm==0.0.17
PyYAML==6.0.2
pyzmq==26.2.1
scikit_learn==1.6.1
sentence_transformers==3.2.1
sentencepiece==0.2.0
spacy==3.8.4
tiktoken==0.8.0
transformers==4.44.2

